hubble: move FakeGRPCServerStream to testutils package
This is useful for any package that needs to implement gRPC tests, thus
move it to a more generic location.
